webscrapers/requirements.txt
selenium >=4.2.0
webdriver-manager >=3.8.5
chromedriver-binary-auto ==0.2.3
flask >=2.2.1
flask_limiter >=3.1.0
flask_cors >=1.10.3